requirements.txt
PyQt6>=6.6.0
markdown>=3.5.1
pymdown-extensions>=10.4
Pygments>=2.17.2
QDarkStyle>=3.2.3